+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- allow using variables for strings of text
+
 ## [0.1.1] - 2020-01-31
 ### Added
 - allow processing of phony recipes with no requirements